refactor some file locations
