+++ src/widget/config/root_path.rs
@@ -5,7 +5,7 @@ use eframe::{
     epaint::{Color32, Vec2},
 };
 
-use crate::{config::Config, font::gen_rich_text, widget::button_icon::ButtonIcon, CLI};
+use crate::{config::Config, library::fonts::gen_rich_text, widget::button_icon::ButtonIcon, CLI};
 
 pub struct ConfigRootPath {
     checkmark: ButtonIcon,
