+++ src/widget/config/video_player.rs
@@ -5,7 +5,9 @@ use eframe::{
     epaint::{Color32, Vec2},
 };
 
-use crate::{config::VideoPlayer, font::gen_rich_text, widget::button_icon::ButtonIcon, CLI};
+use crate::{
+    config::VideoPlayer, library::fonts::gen_rich_text, widget::button_icon::ButtonIcon, CLI,
+};
 
 pub struct ConfigVideoPlayer {
     checkmark: ButtonIcon,
